// FRONTEND/chatbot-admin-panel/src/services/api/schemas/documentSourceUpdate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentSourceUpdateName } from './documentSourceUpdateName';
import type { DocumentSourceUpdateDescription } from './documentSourceUpdateDescription';
import type { DocumentSourceUpdateSourceType } from './documentSourceUpdateSourceType';
import type { DocumentSourceUpdatePathOrConfig } from './documentSourceUpdatePathOrConfig';
import type { DocumentSourceUpdateCredentialsInfo } from './documentSourceUpdateCredentialsInfo';
import type { DocumentSourceUpdateIsActive } from './documentSourceUpdateIsActive';
import type { DocumentSourceUpdateSyncFrequencyCron } from './documentSourceUpdateSyncFrequencyCron';

export interface DocumentSourceUpdate {
  name?: DocumentSourceUpdateName;
  description?: DocumentSourceUpdateDescription;
  source_type?: DocumentSourceUpdateSourceType;
  path_or_config?: DocumentSourceUpdatePathOrConfig;
  credentials_info?: DocumentSourceUpdateCredentialsInfo;
  is_active?: DocumentSourceUpdateIsActive;
  sync_frequency_cron?: DocumentSourceUpdateSyncFrequencyCron;
}