// FRONTEND/chatbot-admin-panel/src/services/api/schemas/documentSourceResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentSourceResponseDescription } from './documentSourceResponseDescription';
import type { DocSourceType } from './docSourceType';
import type { DocumentSourceResponsePathOrConfig } from './documentSourceResponsePathOrConfig';
import type { DocumentSourceResponseSyncFrequencyCron } from './documentSourceResponseSyncFrequencyCron';
import type { DocumentSourceResponseLastSyncedAt } from './documentSourceResponseLastSyncedAt';

export interface DocumentSourceResponse {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  description?: DocumentSourceResponseDescription;
  source_type: DocSourceType;
  /** Ruta (string) o configuración (JSON object) según el source_type. */
  path_or_config: DocumentSourceResponsePathOrConfig;
  is_active?: boolean;
  sync_frequency_cron?: DocumentSourceResponseSyncFrequencyCron;
  id: number;
  last_synced_at?: DocumentSourceResponseLastSyncedAt;
  created_at: string;
  updated_at: string;
}