// FRONTEND/chatbot-admin-panel/src/services/api/schemas/docSourceType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */

export type DocSourceType = typeof DocSourceType[keyof typeof DocSourceType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocSourceType = {
  LOCAL_FOLDER: 'LOCAL_FOLDER',
  S3_BUCKET: 'S3_BUCKET',
  AZURE_BLOB: 'AZURE_BLOB',
  WEB_URL_SINGLE: 'WEB_URL_SINGLE',
} as const;